[Chore] Upgrade dependencies (#159)
* Upgrade to Phoenix 1.6
* Upgrade Elixir to 1.13.1
This fixes a bug that occured when a document didn't contain a description and could have lead to an infinite loop.
This is the last version that is available of the image that we are currently using in our Github Actions pipeline. We'll switch away from this in a separate commit.