Introduce rendering ideas, random path fixes
I'm not 100% sure if taking out ideas was a mistake or if it was
meant as such so I'm not pushing forward with making more of them
visible as even more had been commented out there.

Removed more warnings, however some I was unsure about.

Left some TODO comments on places in the code where I gained
knowledge/know of work to do.