Remove unnecesary relationship declarations in schemas
* the name of the foreign key is usually infered based on naming
* the type is already per default defined in the `Mindwendel.Schema`
* I think it was a mixture of not migrating old code to match the
new after it was changed

On a side note, can't be applied to the join table as
`Mindwendel.Schema` comes with `@derive {Phoenix.Param, key: :id}`
where I'm not sure that's a good default/needed for all the
schemas and hence might be better off being selectively added.

However, I don't know the code well enough to make that call, so
I left it as is.