Move where the feature flag is checked for random inspiration
This one might feel odd. Well, it just felt odd to me that the
function would be called and then just return `nil` if the
feature flag was deactivated. I felt the function shouldn't
be called at all.

This way, for instance, if it was deactivated in the test
environment we could still test it etc - the behavior is still
there.

That's all debatable and more of a "I feel like this" kind of
thing.